Refine existing GroupController tests
